move Pdf.tsx and post.ts, fix code/img, add code for upload
