+++ api/pdf/post.ts
@@ -1,5 +1,6 @@
 import createAPIGatewayProxyHandler from "package/backend/createAPIGatewayProxyHandler";
-import { createZip } from "./Pdf";
+import { createZip } from "../../app/data/Pdf";
+// import uploadFileContent from "package/backend/uploadFileContent";
 
 type Params = {
   files: [
@@ -8,20 +9,29 @@ type Params = {
       content: string;
     }
   ];
+  filename: string;
 };
 
-export const logic = async ({ files }: Params) => {
+export const logic = async ({ files, filename }: Params) => {
   try {
     const zip = await createZip(files);
     const zipBuffer = await zip.generateAsync({ type: "nodebuffer" });
     const zipString = zipBuffer.toString("base64");
     if (!zipString) throw new Error("Failed to create zip");
 
+    // const path = `data/pdfs/${filename}.zip`;
+    // const uploadResult = await uploadFileContent({
+    //   Body: zipBuffer,
+    //   Key: path,
+    // });
+    // if (!uploadResult) throw new Error("Failed to upload zip file");
+
     return {
       statusCode: 200,
       headers: {
         "Content-Type": "application/zip",
       },
+      // body: JSON.stringify(path),
       body: zipString,
       isBase64Encoded: true,
     };
